+++ cde/programs/dtinfo/dtinfo/src/Basic/StringHash.C
@@ -79,8 +79,9 @@ string_hash (const char *key)
   register u_int pos, sum = 0;
   register char c;
 
-  for (pos = 0; c = *key++; pos++)
-    sum = (sum << 5) + (sum >> (sizeof(int) * 8 - 6)) + (c ^ lookup_table[c]);
+  for (pos = 0; (c = *key++); pos++)
+    sum = (sum << 5) + (sum >> (sizeof(int) * 8 - 6))
+	  + (c ^ lookup_table[(int)c]);
 
   // was:
   // sum = c ^ lookup_table [(pos * c) % 256];
